requirements.txt
pyramid
pyramid_jinja2
pyramid_debugtoolbar
pyramid_tm
SQLAlchemy
transaction
zope.sqlalchemy
waitress
psycopg2
nose
rednose
flake8
coverage
python-social-auth
